+++ src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module silly.bean.main {
+	exports io.github.astrapi69.annotation;
+	exports io.github.astrapi69.bean.enumtype;
+	exports io.github.astrapi69.collection.comparator;
+	exports io.github.astrapi69.collection.pair;
+}